apps/mcp_platform/core/requirements.txt
# Core dependencies
kailash-sdk>=0.6.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Async and networking
aiohttp>=3.8.0
websockets>=11.0
httpx>=0.24.0

# Database and caching
redis>=4.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pyotp>=2.8.0
cryptography>=41.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
psutil>=5.9.0
tenacity>=8.2.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing (optional, install with pip install -r requirements-test.txt)
# pytest>=7.3.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
